+++ CHANGELOG.md
@@ -24,7 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Filter parameter on Get-TioExportAsset was a [string], is now a [psobject]
 - Filter parameter on Get-TioExportVuln was a [string], is now a [psobject]
-- Removed incorrect Filter parameter from Get-TioAssetExportStatus call to Invoke-TioApiRequest
+- Removed incorrect Filter parameter from Get-TioExportAssetStatus call to Invoke-TioApiRequest
+- Removed incorrect Filter parameter from Get-TioExportVulnStatus call to Invoke-TioApiRequest
+- Removed incorrect Filter parameter from Get-TioExportAssetChunk call to Invoke-TioApiRequest
+- Removed incorrect Filter parameter from Get-TioExportVulnChunk call to Invoke-TioApiRequest
 
 ## [1.2.0] - 2024-10-15